Task-9: Create a login page where user can access this application. Create a home page where it having facility to manage department and employee details.

+++ mywebapps/crudwebapp/views.py
@@ -1,5 +1,5 @@
-from django.shortcuts import render, redirect
-from django.http import HttpResponse
+from django.shortcuts import render, redirect, render_to_response
+from django.contrib.auth import authenticate, login, logout
 from .models import Department, Employee, GenderChoice
 from .forms import DepartmentForm, EmployeeForm
 
@@ -10,7 +10,9 @@ def index(request):
     :param request: Obj - Request.
     :return: str - HttpResponse response.
     """
-    return HttpResponse("Hello, world. You're at the index.")
+    if is_authenticated(request):
+        return redirect('/home')
+    return render(request, 'login.html')
 
 
 def departments(request):
@@ -18,6 +20,8 @@ def departments(request):
     :param request: Obj - Request.
     :return: Render to department view.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department_list = Department.objects.all()
     return render(
         request, 'departments.html', {'departments': department_list})
@@ -28,6 +32,8 @@ def add_department(request):
     :param request: Obj - Request.
     :return: Renders to departments view on saved or add department page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     if request.method == 'POST':
         form = DepartmentForm(request.POST)
         if form.is_valid():
@@ -48,6 +54,8 @@ def edit_department(request, department_id):
     :param department_id: int - Unique identification of department.
     :return: Render to edit department page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department = Department.objects.get(department_id=department_id)
     return render(
         request, 'department_edit.html', {'department': department})
@@ -59,6 +67,8 @@ def update_department(request, department_id):
     :param department_id: int - Unique identification of department.
     :return: Renders to departments view on update or edit department page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department = Department.objects.get(department_id=department_id)
     form = DepartmentForm(request.POST, instance=department)
     if form.is_valid():
@@ -74,6 +84,8 @@ def delete_department(request, department_id):
     :param department_id: int - Unique identification of department.
     :return: Renders to departments view on delete.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department = Department.objects.get(department_id=department_id)
     department.delete()
     return redirect('/departments')
@@ -84,6 +96,8 @@ def employees(request):
     :param request: Obj - Request.
     :return: Render to employee view.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     employee_list = Employee.objects.all()
     return render(
         request, 'employees.html', {'employees': employee_list})
@@ -94,6 +108,8 @@ def add_employee(request):
     :param request: Obj - Request.
     :return: Renders to employees view on saved or add employee page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department_list = Department.objects.all()
     if request.method == 'POST':
         form = EmployeeForm(request.POST)
@@ -117,6 +133,8 @@ def edit_employee(request, employee_id):
     :param employee_id: int - Unique identification of employee.
     :return: Render to edit employee page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department_list = Department.objects.all()
     employee = Employee.objects.get(employee_id=employee_id)
     return render(
@@ -133,6 +151,8 @@ def update_employee(request, employee_id):
     :param employee_id: int - Unique identification of employee.
     :return: Renders to employees view on update or edit employee page.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     department_list = Department.objects.all()
     employee = Employee.objects.get(employee_id=employee_id)
     form = EmployeeForm(request.POST, instance=employee)
@@ -153,6 +173,63 @@ def delete_employee(request, employee_id):
     :param employee_id: int - Unique identification of department.
     :return: Renders to employees view on delete.
     """
+    if not is_authenticated(request):
+        return redirect('/login')
     employee = Employee.objects.get(employee_id=employee_id)
     employee.delete()
     return redirect('/employees')
+
+
+def authenticate_user(request):
+    """Authenticate user.
+    :param request: Obj - Request.
+    :return: Renders to home page otherwise login page.
+    """
+    if is_authenticated(request):
+        return redirect('/home')
+    username = password = ''
+    if request.POST:
+        username = request.POST['username']
+        password = request.POST['password']
+
+        user = authenticate(username=username, password=password)
+        if user is not None and user.is_active:
+            # login(request, user)
+            request.session['member_id'] = username
+            return redirect('/home')
+    return render(request, 'login.html')
+
+
+def log_out(request):
+    """Logged out and terminate the session of user.
+    :param request: Obj - Request.
+    :return: Renders to logout page.
+    """
+    # logout(request)
+    try:
+        del request.session['member_id']
+    except KeyError:
+        pass
+    return render_to_response('logout.html')
+
+
+def home_page(request):
+    """Redirects to home page.
+    :param request: Obj - Request.
+    :return: Renders to the home page.
+    """
+    if not is_authenticated(request):
+        return redirect('/login')
+    return render(request, 'home.html')
+
+
+def is_authenticated(request):
+    """Check user session in exists.
+    :param request: Obj - Request.
+    :return: bool - True if session exists else False.
+    """
+    try:
+        request.session['member_id']
+    except KeyError:
+        return False
+    return True
